changed structure of components and layouts for desktop/mobile
